+++ CHANGELOG.md
@@ -11,6 +11,23 @@ The format is based on [Keep a Changelog], and this project adheres to
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 [bc]: https://github.com/dogmatiq/.github/blob/main/VERSIONING.md#changelogs
 
+## [Unreleased]
+
+### Added
+
+- Added `envelopepb.Header`, `Body`, `Source`, `Message`, `Extensions`, and
+  `MultiEnvelope`.
+- Added `envelopepb.Packer.CausedBy()` and `envelopepb.MultiPacker` for
+  incrementally building `MultiEnvelope` values.
+
+### Changed
+
+- **[BC]** Changed `envelopepb.Envelope` to use nested `Header` and `Body`
+  messages.
+- **[BC]** Changed `envelopepb.PackOption` from a function type to an
+  interface, and split option roles into `envelopepb.SourcePackOption` and
+  `envelopepb.BodyPackOption`.
+
 ## [0.20.1] - 2026-04-06
 
 ### Added